+++ R/refresh.R
@@ -9,28 +9,34 @@ refresh <- function(x) {
 #' @export
 refresh.data.frame <- function(x) {
   if(!nrow(x)) return(x)
+  # to avoid weird issue with bind_rows
+  if(anyDuplicated(names(x))) return(x)
+
+
   call <- sys.call()
   pf <- parent.frame()
   unrefreshed <- sapply(x, inherits, "reactive_col")
   unrefreshed_vars <- lapply(x[unrefreshed], function(x) {
-    all.vars(attr(x,"reactible_col_def"))
+    all.vars(attr(x,"reactibble_col_def")$expr)
   })
   while(any(unrefreshed)) {
     unrefreshed_bkp <- unrefreshed
     for(var in names(unrefreshed_vars)) {
       dependencies <-unrefreshed_vars[[var]]
       if(!any(na.omit(unrefreshed[dependencies]))){
-        expr <- attr(x[[var]],"reactible_col_def")
-        x[[var]] <- tryCatch(rlang::eval_tidy(expr, x), error = function(e) {
-          missing_vars <- setdiff(all.vars(expr), names(x))
-          msg <- paste0(
-            e$message,
-            "\nDid you drop a necessary variable or provide an incorrect expression?")
-          e$message <- msg
-          e$call <- call
-          stop(e)
-        })
-        x[[var]] <- as_reactive_col(x[[var]], expr)
+        col_def <- attr(x[[var]],"reactibble_col_def")
+        x[[var]] <- tryCatch(
+          eval(col_def$expr, x, col_def$env),
+          error = function(e) {
+            missing_vars <- setdiff(all.vars(col_def), names(x))
+            msg <- paste0(
+              e$message,
+              "\nDid you drop a necessary variable or provide an incorrect expression?")
+            e$message <- msg
+            e$call <- call
+            stop(e)
+          })
+        x[[var]] <- reactive_col(x[[var]], col_def)
         unrefreshed[var] <- FALSE
         unrefreshed_vars[var] <- NULL
       }