+++ NEWS.md
@@ -21,6 +21,6 @@
   columns as "unsynced" colored in red.
 * Optimize process_reactive_dots so reactive columns are coomputed only in the end
 * Use quosures instead of expressions as column definitions
-* implement temporary (brittle) solution to rbind and bind_rows
+* implement temporary solution to rbind and bind_rows using vctrs